added forgot and reset password features, organized code better.
