+++ package.json
@@ -15,8 +15,11 @@
     "bcryptjs": "^2.4.3",
     "dotenv": "^16.4.7",
     "express": "^4.21.2",
+    "express-async-errors": "^3.1.1",
+    "handlebars": "^4.7.8",
     "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.9.4",
+    "nodemailer": "^6.9.16",
     "validator": "^13.12.0"
   },
   "devDependencies": {
